# lib/api-zod/src/generated/types/acceptInviteResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */
import type { AcceptInviteResultStatus } from './acceptInviteResultStatus';

export interface AcceptInviteResult {
  ownerId: string;
  status: AcceptInviteResultStatus;
}