# lib/api-zod/src/generated/types/companyCollaborator.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */

/**
 * Pihak yang dapat mengakses sebuah perusahaan (pemilik atau anggota berbagi).
 */
export interface CompanyCollaborator {
  userId: string;
  /** Email anggota atau label pemilik */
  name: string;
  isOwner: boolean;
  /**
     * Id baris berbagi (null untuk pemilik)
     * @nullable
     */
  shareId?: number | null;
}