# lib/api-zod/src/generated/types/shareCompanyInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */

export interface ShareCompanyInput {
  /**
     * Clerk userId anggota tim aktif yang diberi akses
     * @minLength 1
     */
  memberId: string;
}