# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * LKPM-Flow API — companies, reports, data points, constraints, audit trail, and dashboards for LKPM (Laporan Kegiatan Penanaman Modal) reporting to OSS.
 * OpenAPI spec version: 0.1.0
 */
import type { TeamMemberStatus } from './teamMemberStatus';

/**
 * Anggota tim yang diundang oleh pemilik (owner). inviteCode hanya dikembalikan kepada pemilik untuk dibagikan ke calon anggota.
 */
export interface TeamMember {
  id: number;
  email: string;
  status: TeamMemberStatus;
  /**
     * Clerk userId anggota; terisi setelah undangan diterima
     * @nullable
     */
  memberId?: string | null;
  /**
     * Kode undangan untuk diserahkan ke calon anggota
     * @nullable
     */
  inviteCode?: string | null;
  createdAt: Date;
}